#fix Formがとりあえず動くように修正 #change UnixTime関連の型をuint64に変更 #change Main関数をtry-catch構文に変更 (これはログが記録されるようにするため) #add サーバー起動用bat

+++ MainProject/KosenProcon30/Game.hpp
@@ -48,7 +48,7 @@ namespace Procon30 {
 		int32 gameID;
 		//0,1,2...
 		int32 gameNum;
-		int32 startedAtUnixTime;
+		uint64 startedAtUnixTime;
 		String matchTo;
 		int32 turnMillis;
 		int32 intervalMillis;
